requirements.txt
absl-py==2.1.0
asciitree==0.3.3
chex==0.1.88
click==8.1.8
cloudpickle==3.1.1
dask==2024.8.0
distributed==2024.8.0
dm-haiku==0.0.13
etils==1.5.2
fasteners==0.19
fsspec==2024.12.0
importlib_metadata==8.6.1
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.5
jmp==0.0.4
jraph==0.0.6.dev0
locket==1.0.0
MarkupSafe==3.0.2
ml_dtypes==0.5.1
msgpack==1.1.0
numcodecs==0.12.1
numpy==2.0.2
opt_einsum==3.4.0
optax==0.2.4
packaging==24.2
pandas==2.2.3
partd==1.4.2
psutil==6.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scipy==1.13.1
six==1.17.0
sortedcontainers==2.4.0
tabulate==0.9.0
tblib==3.0.0
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
xarray==2024.7.0
zarr==2.18.2
zict==3.0.0
zipp==3.21.0
